refactor :: Keyboard height 오류 해결
